+++ App4/model/User.cs
@@ -19,7 +19,7 @@ namespace App4.Model
         private string Topic = "Tourist";
         public int id { get; set; }
         public string name { get; set; }
-        public int identifier { get; set; }
+        public double identifier { get; set; }
         public string profilePic { get; set; }
         public int status { get; set; }
         public int wristbandID { get; set; }
@@ -37,7 +37,7 @@ namespace App4.Model
             this.bitmap = null;
         }
 
-        public User(int id, string name, int identifier, string profilePic, int status, int wristbandID)
+        public User(int id, string name, double identifier, string profilePic, int status, int wristbandID)
         {
             this.id = id;
             this.name = name;
@@ -47,7 +47,7 @@ namespace App4.Model
             this.wristbandID = wristbandID;
         }
 
-        public User(int id, string name, int identifier, string profilePic, int status, int wristbandID, BitmapImage bitmap)
+        public User(int id, string name, double identifier, string profilePic, int status, int wristbandID, BitmapImage bitmap)
         {
             this.id = id;
             this.name = name;
@@ -57,7 +57,7 @@ namespace App4.Model
             this.wristbandID = wristbandID;
             this.bitmap = bitmap;
         }
-        public User(string name, int identifier, string profilePic, int wristbandID)
+        public User(string name, double identifier, string profilePic, int wristbandID)
         {
             this.name = name;
             this.identifier = identifier;
@@ -163,7 +163,7 @@ namespace App4.Model
                 {
                     int id = -1;
                     string name = "";
-                    int identifier = -1;
+                    double identifier = -1;
                     string profilePic = "";
                     int status = -1;
                     int wristbandID = -1;
@@ -172,7 +172,7 @@ namespace App4.Model
                     if (jsonArr.GetObjectAt(i)["name"].ValueType != JsonValueType.Null)
                         name = jsonArr.GetObjectAt(i).GetNamedString("name");
                     if (jsonArr.GetObjectAt(i)["identifier"].ValueType != JsonValueType.Null)
-                        identifier = Int32.Parse(jsonArr.GetObjectAt(i).GetNamedString("identifier"));
+                        identifier = Double.Parse(jsonArr.GetObjectAt(i).GetNamedString("identifier"));
                     if (jsonArr.GetObjectAt(i)["profile_pic"].ValueType != JsonValueType.Null)
                         profilePic = jsonArr.GetObjectAt(i).GetNamedString("profile_pic");
                     if (jsonArr.GetObjectAt(i)["status"].ValueType != JsonValueType.Null)
@@ -204,7 +204,7 @@ namespace App4.Model
             JsonValue json = JsonValue.Parse(response.Content.ReadAsStringAsync().Result.ToString());
             int id_pri = -1;
             string name = "";
-            int identifier = -1;
+            double identifier = -1;
             string profilePic = "";
             int status = -1;
             int wristbandID = -1;
@@ -213,7 +213,7 @@ namespace App4.Model
             if (json.GetObject()["name"].ValueType != JsonValueType.Null)
                 name = json.GetObject().GetNamedString("name");
             if (json.GetObject()["identifier"].ValueType != JsonValueType.Null)
-                identifier = Int32.Parse(json.GetObject().GetNamedString("identifier"));
+                identifier = Double.Parse(json.GetObject().GetNamedString("identifier"));
             if (json.GetObject()["profile_pic"].ValueType != JsonValueType.Null)
                 profilePic = json.GetObject().GetNamedString("profile_pic");
             if (json.GetObject()["status"].ValueType != JsonValueType.Null)
